+++ src/aks-preview/HISTORY.md
@@ -2,9 +2,15 @@
 
 Release History
 ===============
-0.5.3
+0.5.4
 +++++
 * Add operations of maintenance configuration
+=======
+
+0.5.3
++++++
+* Add `--enable-pod-identity-with-kubenet` for enabling AAD Pod Identity in Kubenet cluster
+* Add `--fqdn-subdomain parameter` to create private cluster with custom private dns zone scenario
 
 0.5.2
 +++++